Refactor: split merge plan class
